Distributed Systems Netbeans Project
This is the finished product of the Distributed Systems Project as outlined by the module. Due to financial circumstances the repository is public. Due to the solo nature of the project there has been no need to push the project to a repository. This has only been completed to satisfy the project specification.